+++ app/src/test/java/com/piotv/keytab/file/FileManagerFragmentTest.kt
@@ -4,6 +4,7 @@ import android.view.View
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.tabs.TabLayout
 import com.piotv.keytab.R
 import com.piotv.keytab.ime.KeyTabExecutors
@@ -14,6 +15,7 @@ import org.junit.runner.RunWith
 import org.robolectric.Robolectric
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.annotation.Config
+import org.robolectric.shadows.ShadowApplication
 import org.robolectric.shadows.ShadowLooper
 import java.io.File
 import java.util.concurrent.TimeUnit